requirements.txt
requests
pytrends
matplotlib
python-dotenv